Partially revert 4c5714eaa6e14294b47a944832da857f36bc6f7b

+++ src/core/FlashAccountingCommon.sol
@@ -657,7 +657,7 @@ library Decoder {
         if (state.globalSell().tokenIsEth()) {
             assert(payer == address(this));
 
-            uint256 bps;
+            uint16 bps;
             assembly ("memory-safe") {
                 // `data` hasn't been advanced from decoding `sellToken` above. so we have to
                 // implicitly advance it by 20 bytes to decode `bps` then advance by 22 bytes
@@ -674,7 +674,7 @@ library Decoder {
             }
         } else {
             if (payer == address(this)) {
-                uint256 bps;
+                uint16 bps;
                 assembly ("memory-safe") {
                     // `data` hasn't been advanced from decoding `sellToken` above. so we have to
                     // implicitly advance it by 20 bytes to decode `bps` then advance by 22 bytes
